fix: 修正文章分区标准化并抽取 OneBot 调用共享类型
